pyproject.toml
[project]
name = "create-knowlege-graph-app"
version = "0.1.0"
description = "Japanese NER + Relation Extraction to Neo4j (Cypher)"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "transformers>=4.42.0",
  "torch>=2.2.0",
  "sentencepiece>=0.1.99",
  "sudachipy>=0.6.8",
  "sudachidict-core>=20240109",
  "neo4j>=5.23.0",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.4",
  "streamlit>=1.36.0",
  "pyvis>=0.3.2",
  "networkx>=3.2.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["create_knowlege_graph_app"]

[tool.setuptools.package-dir]
"" = "src"


